// types/product.ts
export interface Product {
    id: number;
    nombre: string;
    precio: number;
    descripcion?: string;
    imageUrl?: string;
    categoryId?: number;
    category?: Category;
    createdAt?: string;
    updatedAt?: string;
}

export interface Category {
    id: number;
    nombre: string;
}

export interface ApiResponse<T> {
    success: boolean;
    message: string;
    data: T;
}